导读/总结工具移到 hamster-reading-mcp（纠正归属）
原计划就是放 reading-mcp（同属 All About Book 阅读域，与书摘/
问答工具住在一起），上一提交误放进了 lounge-mcp。本次整体平移：

- 8 个导读/总结工具原样迁入 hamster-reading-mcp（10 → 18 个工具）
- hamster-lounge-mcp 恢复原状（14 个工具，不再引用 AAB client）
- README 工具表同步；两个函数均已重新部署并验证
